+++ lib/reminders_list.dart
@@ -2,7 +2,7 @@ class RemList {
   String title;
   String id;
 
-  RemList(this.title, this.id);
+  RemList(this.title, {this.id = ''});
 
   RemList.fromJson(json)
       : title = json['title'],