+++ CHANGELOG.md
@@ -1,3 +1,24 @@
-## 0.0.1
+## 1.0.0 2022-11-05
 
-* TODO: Describe initial release.
+* Add saveReminder to create reminders and save changes
+* Full version
+  
+## 0.1.0 2020-12-16
+
+* Update README with correct Key Value pair in info.plist
+* Change Reminders API to an object from a static class
+* Tidy API method names
+
+## 0.0.2 2020-11-30
+
+* Add API documentation
+* Code tidy
+* Dart package does not rely on Flutter
+
+## 0.0.1 2020-11-30
+
+* Initial release
+* Get default reminder list
+* Get list of reminder list names
+* Get reminders in a list by list name
+* Delete a reminder by ID